Completion
Completion

+++ application/controllers/Client.php
@@ -31,8 +31,7 @@ class Client extends CI_Controller {
         }
         $this->load->view('server-page', $data);
     }
-   
- 
+
     public function GUID() {
         if (function_exists('com_create_guid') === true) {
             return trim(com_create_guid(), '{}');
@@ -40,25 +39,32 @@ class Client extends CI_Controller {
 
         return sprintf('%04X%04X-%04X-%04X-%04X-%04X%04X%04X', mt_rand(0, 65535), mt_rand(0, 65535), mt_rand(0, 65535), mt_rand(16384, 20479), mt_rand(32768, 49151), mt_rand(0, 65535), mt_rand(0, 65535), mt_rand(0, 65535));
     }
-  
+
     public function delete() {
-        $this->load->helper(array('form', 'url'));              
-        $id = urldecode($this->uri->segment(3));        
-        $name = urldecode($this->uri->segment(4));
-        $query = $this->Md->cascade($name,'sync_data','client');
-         $query = $this->Md->delete($id, 'client');
-        //cascade($id,$table,$field)
-        
-       if ($this->db->affected_rows() > 0) {
-           
-         $this->session->set_flashdata('msg', '<div class="alert alert-error">
+        if ($this->session->userdata('level') == 1) {
+            $this->load->helper(array('form', 'url'));
+            $id = urldecode($this->uri->segment(3));
+            $name = urldecode($this->uri->segment(4));
+            $query = $this->Md->cascade($name, 'sync_data', 'client');
+            $query = $this->Md->delete($id, 'client');
+            //cascade($id,$table,$field)
+
+            if ($this->db->affected_rows() > 0) {
+
+                $this->session->set_flashdata('msg', '<div class="alert alert-error">
                                                    
                                                 <strong>
                                                 Information deleted	</strong>									
 						</div>');
+                redirect('client', 'refresh');
+            }
+        } else {
+            $this->session->set_flashdata('msg', '<div class="alert alert-error">                                                   
+                                                <strong>
+                                                 You cannot carry out this action ' . '	</strong>									
+						</div>');
             redirect('client', 'refresh');
-       }
-        
-    }    
+        }
+    }
 
 }
